Moved to traversal.py
