+++ sympy/core/__init__.py
@@ -4,7 +4,7 @@
 from .sympify import sympify, SympifyError
 from .cache import cacheit
 from .assumptions import assumptions, check_assumptions, failing_assumptions, common_assumptions
-from .basic import Basic, Atom, preorder_traversal, bottom_up, use
+from .basic import Basic, Atom
 from .singleton import S
 from .expr import Expr, AtomicExpr, UnevaluatedExpr
 from .symbol import Symbol, Wild, Dummy, symbols, var
@@ -29,6 +29,7 @@ from .containers import Tuple, Dict
 from .exprtools import gcd_terms, factor_terms, factor_nc
 from .parameters import evaluate
 from .kind import UndefinedKind, NumberKind, BooleanKind
+from .traversal import preorder_traversal, bottom_up, use
 
 # expose singletons
 Catalan = S.Catalan
@@ -44,7 +45,7 @@ __all__ = [
     'assumptions', 'check_assumptions', 'failing_assumptions',
     'common_assumptions',
 
-    'Basic', 'Atom', 'preorder_traversal', 'bottom_up', 'use',
+    'Basic', 'Atom',
 
     'S',
 
@@ -91,4 +92,6 @@ __all__ = [
     'TribonacciConstant',
 
     'UndefinedKind', 'NumberKind', 'BooleanKind',
+
+    'preorder_traversal', 'bottom_up', 'use',
 ]
