+++ sympy/core/basic.py
@@ -2004,107 +2004,6 @@ def _atomic(e, recursive=False):
     return atoms
 
 
-class preorder_traversal:
-    """
-    Do a pre-order traversal of a tree.
-
-    This iterator recursively yields nodes that it has visited in a pre-order
-    fashion. That is, it yields the current node then descends through the
-    tree breadth-first to yield all of a node's children's pre-order
-    traversal.
-
-
-    For an expression, the order of the traversal depends on the order of
-    .args, which in many cases can be arbitrary.
-
-    Parameters
-    ==========
-    node : sympy expression
-        The expression to traverse.
-    keys : (default None) sort key(s)
-        The key(s) used to sort args of Basic objects. When None, args of Basic
-        objects are processed in arbitrary order. If key is defined, it will
-        be passed along to ordered() as the only key(s) to use to sort the
-        arguments; if ``key`` is simply True then the default keys of ordered
-        will be used.
-
-    Yields
-    ======
-    subtree : sympy expression
-        All of the subtrees in the tree.
-
-    Examples
-    ========
-
-    >>> from sympy import symbols
-    >>> from sympy.core.basic import preorder_traversal
-    >>> x, y, z = symbols('x y z')
-
-    The nodes are returned in the order that they are encountered unless key
-    is given; simply passing key=True will guarantee that the traversal is
-    unique.
-
-    >>> list(preorder_traversal((x + y)*z, keys=None)) # doctest: +SKIP
-    [z*(x + y), z, x + y, y, x]
-    >>> list(preorder_traversal((x + y)*z, keys=True))
-    [z*(x + y), z, x + y, x, y]
-
-    """
-    def __init__(self, node, keys=None):
-        self._skip_flag = False
-        self._pt = self._preorder_traversal(node, keys)
-
-    def _preorder_traversal(self, node, keys):
-        yield node
-        if self._skip_flag:
-            self._skip_flag = False
-            return
-        if isinstance(node, Basic):
-            if not keys and hasattr(node, '_argset'):
-                # LatticeOp keeps args as a set. We should use this if we
-                # don't care about the order, to prevent unnecessary sorting.
-                args = node._argset
-            else:
-                args = node.args
-            if keys:
-                if keys != True:
-                    args = ordered(args, keys, default=False)
-                else:
-                    args = ordered(args)
-            for arg in args:
-                yield from self._preorder_traversal(arg, keys)
-        elif iterable(node):
-            for item in node:
-                yield from self._preorder_traversal(item, keys)
-
-    def skip(self):
-        """
-        Skip yielding current node's (last yielded node's) subtrees.
-
-        Examples
-        ========
-
-        >>> from sympy.core import symbols
-        >>> from sympy.core.basic import preorder_traversal
-        >>> x, y, z = symbols('x y z')
-        >>> pt = preorder_traversal((x+y*z)*z)
-        >>> for i in pt:
-        ...     print(i)
-        ...     if i == x+y*z:
-        ...             pt.skip()
-        z*(x + y*z)
-        z
-        x + y*z
-        """
-        self._skip_flag = True
-
-    def __next__(self):
-        return next(self._pt)
-
-    def __iter__(self):
-        return self
-
-
 def _make_find_query(query):
     """Convert the argument of Basic.find() into a callable"""
     try:
@@ -2118,99 +2017,6 @@ def _make_find_query(query):
     return query
 
 
-def use(expr, func, level=0, args=(), kwargs={}):
-    """
-    Use ``func`` to transform ``expr`` at the given level.
-
-    Examples
-    ========
-
-    >>> from sympy import use, expand
-    >>> from sympy.abc import x, y
-
-    >>> f = (x + y)**2*x + 1
-
-    >>> use(f, expand, level=2)
-    x*(x**2 + 2*x*y + y**2) + 1
-    >>> expand(f)
-    x**3 + 2*x**2*y + x*y**2 + 1
-
-    """
-    def _use(expr, level):
-        if not level:
-            return func(expr, *args, **kwargs)
-        else:
-            if expr.is_Atom:
-                return expr
-            else:
-                level -= 1
-                _args = []
-
-                for arg in expr.args:
-                    _args.append(_use(arg, level))
-
-                return expr.__class__(*_args)
-
-    return _use(sympify(expr), level)
-
-
-def walk(e, *target):
-    """Iterate through the args that are the given types (target) and
-    return a list of the args that were traversed; arguments
-    that are not of the specified types are not traversed.
-
-    Examples
-    ========
-
-    >>> from sympy.simplify.simplify import walk
-    >>> from sympy import Min, Max
-    >>> from sympy.abc import x, y, z
-    >>> list(walk(Min(x, Max(y, Min(1, z))), Min))
-    [Min(x, Max(y, Min(1, z)))]
-    >>> list(walk(Min(x, Max(y, Min(1, z))), Min, Max))
-    [Min(x, Max(y, Min(1, z))), Max(y, Min(1, z)), Min(1, z)]
-
-    See Also
-    ========
-
-    bottom_up
-    """
-    if isinstance(e, target):
-        yield e
-        for i in e.args:
-            yield from walk(i, *target)
-
-
-def bottom_up(rv, F, atoms=False, nonbasic=False, simultaneous=True):
-    """Apply ``F`` to all expressions in an expression tree from the
-    bottom up. If ``atoms`` is True, apply ``F`` even if there are no args;
-    if ``nonbasic`` is True, try to apply ``F`` to non-Basic objects.
-    """
-    args = getattr(rv, 'args', None)
-    if args is not None:
-        if args:
-            newargs = tuple([bottom_up(a, F, atoms, nonbasic, simultaneous) for a in args])
-            if args != newargs:
-                rv = rv.func(*newargs)
-                if simultaneous:
-                    # if rv is something that was already
-                    # matched (that was changed) then skip
-                    # applying F again
-                    for i, e in enumerate(args):
-                        if rv == e and e != newargs[i]:
-                            return rv
-            rv = F(rv)
-        elif atoms:
-            rv = F(rv)
-    else:
-        if nonbasic:
-            try:
-                rv = F(rv)
-            except TypeError:
-                pass
-
-    return rv
-
-
 # Delayed to avoid cyclic import
 from .singleton import S
+from .traversal import bottom_up, preorder_traversal
