+++ sympy/functions/elementary/miscellaneous.py
@@ -1,7 +1,6 @@
 from sympy.core import Function, S, sympify
 from sympy.utilities.iterables import sift
 from sympy.core.add import Add
-from sympy.core.basic import walk
 from sympy.core.containers import Tuple
 from sympy.core.compatibility import ordered
 from sympy.core.operations import LatticeOp, ShortCircuit
@@ -17,6 +16,7 @@ from sympy.core.singleton import Singleton
 from sympy.core.symbol import Dummy
 from sympy.core.rules import Transform
 from sympy.core.logic import fuzzy_and, fuzzy_or, _torf
+from sympy.core.traversal import walk
 from sympy.logic.boolalg import And, Or
