+++ sympy/simplify/cse_opts.py
@@ -2,8 +2,8 @@
 opportunities.
 """
 from sympy.core import Add, Basic, Mul
-from sympy.core.basic import preorder_traversal
 from sympy.core.singleton import S
+from sympy.core.traversal import preorder_traversal
 from sympy.utilities.iterables import default_sort_key
